refactor(WinningNumbersTest): 당첨 번호와 관련된 테스트를 LottoTest에서 WinningNumbersTest로 이동
