+++ src/test/java/lotto/WinningNumbersTest.java
@@ -1,0 +1,39 @@
+package lotto;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import java.util.List;
+import lotto.model.WinningNumbers;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class WinningNumbersTest {
+
+    @DisplayName("당첨 번호에 중복된 숫자가 있으면 예외가 발생한다.")
+    @Test
+    void 중복_당첨_번호_예외_테스트() {
+        assertThatThrownBy(() -> new WinningNumbers(List.of(1, 2, 3, 4, 5, 5)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("당첨 번호의 개수가 6보다 작으면 예외가 발생한다.")
+    @Test
+    void 적은_당첨_번호_개수_예외_테스트() {
+        assertThatThrownBy(() -> new WinningNumbers(List.of(1, 2, 3, 4, 5)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("당첨 번호의 개수가 6보다 크면 예외가 발생한다.")
+    @Test
+    void 많은_당첨_번호_개수_예외_테스트() {
+        assertThatThrownBy(() -> new WinningNumbers(List.of(1, 2, 3, 4, 5, 6, 7)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("당첨 번호의 범위가 1 이상 45 이하가 아니면 예외가 발생한다.")
+    @Test
+    void 당첨_번호_범위_예외_테스트() {
+        assertThatThrownBy(() -> new WinningNumbers(List.of(1, 2, 3, 4, 5, 47)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+}
